Moved Windows build to Visual Studio 2019 x64

+++ src/Vulkan/Window.cpp
@@ -101,7 +101,7 @@ std::vector<const char*> Window::GetRequiredInstanceExtensions() const
 	return std::vector<const char*>(glfwExtensions, glfwExtensions + glfwExtensionCount);
 }
 
-double Window::ContentScale() const
+float Window::ContentScale() const
 {
 	float xscale;
 	float yscale;
